Перенос логики из MainActivity в EditActivity
