+++ app/src/main/java/com/example/carstz/Car.kt
@@ -1,3 +1,3 @@
 package com.example.carstz
 
-data class Car(val imageId: Int, val title: String)
+data class Car(val imageId: Int, val title: String, val desc: String)
