+++ app/src/main/java/com/example/carstz/MainActivity.kt
@@ -9,18 +9,6 @@ import com.example.carstz.databinding.ActivityMainBinding
 class MainActivity : AppCompatActivity() {
     lateinit var binding: ActivityMainBinding
     private val adapter = CarAdapter()
-    private val imageIdList = listOf(
-        R.drawable.audia4,
-        R.drawable.audia6,
-        R.drawable.audia8,
-        R.drawable.bmwx3,
-        R.drawable.bmwx5,
-        R.drawable.bmwx6,
-        R.drawable.volkswagengolf,
-        R.drawable.volkswagenpassat,
-        R.drawable.volkswagenpolo
-    )
-    private var index = 0
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -33,10 +21,7 @@ class MainActivity : AppCompatActivity() {
         rcView.layoutManager = GridLayoutManager(this@MainActivity, 3)
         rcView.adapter = adapter
         buttonAdd.setOnClickListener {
-            if (index > 8) index = 0
-            val car = Car(imageIdList[index],"Car $index")
-            adapter.addCar(car)
-            index++
+
         }
     }
 }
